Moved common code to DCommon, include dir builds working in dmd too.
